requirements.txt
flask
gunicorn
werkzeug
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-groq
python-dotenv
flask-cors
faiss-cpu
sentence-transformers